+++ src/problem/Problems.kt
@@ -5,6 +5,7 @@ import discrete.Problem
 const val BinCover = "bin_cover"
 const val BinPacking = "bin_packing"
 const val Clique = "clique"
+const val CliqueCover = "clique_cover"
 const val Interval = "interval"
 const val KCut = "k_cut"
 const val Knapsack = "knapsack"
@@ -24,6 +25,7 @@ val Creator: Map<String, (String, Int) -> Problem?>  = mapOf(
 	BinCover to ::newBinCover,
 	BinPacking to ::newBinPacking,
 	Clique to ::newClique,
+	CliqueCover to ::newCliqueCover,
 	Interval to ::newInterval,
 	KCut to ::newKCut,
 	Knapsack to ::newKnapsack,